requirements.txt
discord.py==2.6.4
python-dotenv==1.2.1
aiohttp==3.12.15